rename ethtool add ethtool_windows
